requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-environ==0.4.5
django-filter==2.3.0
django-heroku==0.3.1
django-phone-field==1.8.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.7.0
jsmin==3.0.1
Markdown==3.2.2
Pillow==9.2.0
psycopg==3.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
zipp==3.1.0
